refactor: get rid of tokens module
