+++ apps/cowswap-frontend/src/common/types.ts
@@ -1,4 +1,4 @@
-import type { AppData, AppDataHash, EnrichedOrder, PriceQuality, SupportedChainId } from '@cowprotocol/cow-sdk'
+import { BalancesState } from '@cowprotocol/balances-and-allowances'
 import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 
 /**
@@ -30,17 +30,7 @@ export interface TradeAmounts {
   readonly outputAmount: CurrencyAmount<Currency>
 }
 
-export interface FeeQuoteParams extends Pick<EnrichedOrder, 'sellToken' | 'buyToken' | 'kind'> {
-  amount: string
-  userAddress?: string | null
-  receiver?: string | null
-  validFor?: number
-  fromDecimals?: number
-  toDecimals?: number
-  chainId: SupportedChainId
-  priceQuality: PriceQuality
-  isBestQuote?: boolean
-  isEthFlow: boolean
-  appData?: AppData
-  appDataHash?: AppDataHash
+export interface BalancesAndAllowances {
+  balances: BalancesState['values']
+  isLoading: boolean
 }
