+++ apps/cowswap-frontend/src/modules/limitOrders/hooks/useTradeFlowContext.ts
@@ -14,11 +14,11 @@ import { useRateImpact } from 'modules/limitOrders/hooks/useRateImpact'
 import { TradeFlowContext } from 'modules/limitOrders/services/types'
 import { limitOrdersSettingsAtom } from 'modules/limitOrders/state/limitOrdersSettingsAtom'
 import { useGeneratePermitHook, useGetCachedPermit, usePermitInfo } from 'modules/permit'
-import { useEnoughBalanceAndAllowance } from 'modules/tokens'
 import { TradeType, useAmountsToSign } from 'modules/trade'
 import { useTradeQuote } from 'modules/tradeQuote'
 
 import { useGP2SettlementContract } from 'common/hooks/useContract'
+import { useEnoughAllowance } from 'common/hooks/useEnoughBalance'
 import { useSafeMemo } from 'common/hooks/useSafeMemo'
 
 import { useLimitOrdersDerivedState } from './useLimitOrdersDerivedState'
@@ -40,9 +40,7 @@ export function useTradeFlowContext(): TradeFlowContext | null {
   const permitInfo = usePermitInfo(state.inputCurrency, TradeType.LIMIT_ORDER)
   const { maximumSendSellAmount } = useAmountsToSign() || {}
 
-  const { enoughAllowance } = useEnoughBalanceAndAllowance({
-    amount: maximumSendSellAmount || undefined,
-  })
+  const enoughAllowance = useEnoughAllowance(maximumSendSellAmount || undefined)
   const generatePermitHook = useGeneratePermitHook()
   const getCachedPermit = useGetCachedPermit()
