+++ apps/cowswap-frontend/src/modules/ordersTable/pure/OrdersTableContainer/index.cosmos.tsx
@@ -3,8 +3,7 @@ import { useEffect } from 'react'
 
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
 
-import { BalancesAndAllowances } from 'modules/tokens'
-
+import { BalancesAndAllowances } from 'common/types'
 import { ParsedOrder } from 'utils/orderUtils/parseOrder'
 
 import { ordersMock } from './orders.mock'
