+++ apps/cowswap-frontend/src/modules/ordersTable/pure/WarningEstimatedPrice/index.tsx
@@ -5,8 +5,7 @@ import { Token } from '@uniswap/sdk-core'
 
 import { OrderStatus } from 'legacy/state/orders/actions'
 
-import type { BalancesAndAllowances } from 'modules/tokens'
-
+import { BalancesAndAllowances } from 'common/types'
 import { ParsedOrder } from 'utils/orderUtils/parseOrder'
 
 import { EstimatedExecutionPrice } from '../../containers/OrderRow/EstimatedExecutionPrice'
@@ -58,15 +57,15 @@ export function WarningEstimatedPrice(props: WarningEstimatedPriceProps) {
 }
 
 // TODO: Reduce function complexity by extracting logic
-// eslint-disable-next-line complexity
+
 const findWarningChildWithParams = ({
   isChild,
   childOrders,
   isTwapTable,
   chainId,
   balancesAndAllowances,
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
+  // TODO: Add proper return type annotation
+  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 }: WarningEstimatedPriceProps) => {
   if (!isTwapTable || isChild || !childOrders) return null
