+++ apps/cowswap-frontend/src/modules/ordersTable/types.ts
@@ -6,9 +6,9 @@ import type { Currency, Price, Token } from '@uniswap/sdk-core'
 import type { Order } from 'legacy/state/orders/actions'
 
 import type { PendingOrdersPrices, SpotPricesKeyParams } from 'modules/orders'
-import type { BalancesAndAllowances } from 'modules/tokens'
 
 import type { UseCancelOrderReturn } from 'common/hooks/useCancelOrder'
+import { BalancesAndAllowances } from 'common/types'
 import type { CancellableOrder } from 'common/utils/isOrderCancellable'
 import type { ParsedOrder } from 'utils/orderUtils/parseOrder'
