+++ apps/cowswap-frontend/src/modules/ordersTable/utils/getOrderParams.ts
@@ -3,9 +3,8 @@ import { SupportedChainId } from '@cowprotocol/cow-sdk'
 import { BigNumber } from '@ethersproject/bignumber'
 import { Currency, CurrencyAmount, Percent, Token } from '@uniswap/sdk-core'
 
-import { BalancesAndAllowances } from 'modules/tokens'
-
 import { RateInfoParams } from 'common/pure/RateInfo'
+import { BalancesAndAllowances } from 'common/types'
 import { getOrderPermitAmount } from 'utils/orderUtils/getOrderPermitAmount'
 import { ParsedOrder } from 'utils/orderUtils/parseOrder'
