--- apps/cowswap-frontend/src/modules/tokens/index.ts
@@ -1,5 +1,0 @@
-export * from './types'
-
-// Exported for all other cases, when we need the effective balance (i.e. )
-export * from './hooks/useEnoughBalance'
-export * from './hooks/useBalancesAndAllowances'
