--- apps/cowswap-frontend/src/modules/tokens/types.ts
@@ -1,6 +1,0 @@
-import { BalancesState } from '@cowprotocol/balances-and-allowances'
-
-export interface BalancesAndAllowances {
-  balances: BalancesState['values']
-  isLoading: boolean
-}
