+++ apps/cowswap-frontend/src/modules/tradeQuote/hooks/useTradeQuotePolling.test.tsx
@@ -10,8 +10,9 @@ import { JotaiTestProvider, WithMockedWeb3 } from 'test-utils'
 import { bridgingSdk } from 'tradingSdk/bridgingSdk'
 
 import { LimitOrdersDerivedState, limitOrdersDerivedStateAtom } from 'modules/limitOrders/state/limitOrdersRawStateAtom'
-import * as tokensModule from 'modules/tokens'
 import { DEFAULT_TRADE_DERIVED_STATE, TradeType } from 'modules/trade'
+
+import { useEnoughAllowance } from 'common/hooks/useEnoughBalance'
 
 import { useTradeQuotePolling } from './useTradeQuotePolling'
 
@@ -20,6 +21,10 @@ import { tradeQuoteInputAtom } from '../state/tradeQuoteInputAtom'
 
 jest.mock('modules/zeroApproval/hooks/useZeroApprovalState')
 jest.mock('common/hooks/useGetMarketDimension')
+jest.mock('common/hooks/useEnoughBalance', () => ({
+  ...jest.requireActual('common/hooks/useEnoughBalance'),
+  useEnoughAllowance: jest.fn().mockReturnValue(undefined),
+}))
 jest.mock('@cowprotocol/common-hooks', () => ({
   ...jest.requireActual('@cowprotocol/common-hooks'),
   useIsWindowVisible: jest.fn().mockReturnValue(true),
@@ -34,7 +39,7 @@ jest.mock('tradingSdk/bridgingSdk', () => ({
     getQuote: jest.fn(),
   },
 }))
-const useEnoughBalanceAndAllowanceMock = jest.spyOn(tokensModule, 'useEnoughBalanceAndAllowance')
+const useEnoughAllowanceMock = useEnoughAllowance as jest.Mock
 
 const bridgingSdkMock = bridgingSdk as unknown as { getQuote: jest.Mock }
 
@@ -83,7 +88,7 @@ describe('useTradeQuotePolling()', () => {
 
     bridgingSdkMock.getQuote.mockImplementation(() => new Promise(() => void 0))
 
-    useEnoughBalanceAndAllowanceMock.mockReturnValue({ enoughBalance: true, enoughAllowance: true })
+    useEnoughAllowanceMock.mockReturnValue(true)
   })
 
   describe('When wallet is connected', () => {
